+++ lib/webspicy/web/specification/post.rb
@@ -1,0 +1,9 @@
+module Webspicy
+  module Web
+    class Specification
+      module Post
+        include Webspicy::Specification::Post
+      end # module Post
+    end # class Specification
+  end # module Web
+end # module Webspicy