+++ lib/webspicy/web/specification/pre.rb
@@ -1,0 +1,9 @@
+module Webspicy
+  module Web
+    class Specification
+      module Pre
+        include Webspicy::Specification::Pre
+      end # module Pre
+    end # class Specification
+  end # module Web
+end # module Webspicy